Split diff viewer row blocks
